feat: added jest,supertest and added node_env check on sync db

+++ express-mysql-omdbapi/src/server.js
@@ -4,9 +4,11 @@ require('dotenv').config()
 const db = require("./models")
 
 const app = express();
-db.sequelize.sync({ force: true }).then(() => {
-  console.log("Drop and re-sync db.");
-});
+if (process.env.NODE_ENV && process.env.NODE_ENV === "DEV") {
+  db.sequelize.sync({ force: true }).then(() => {
+    console.log("Drop and re-sync db.");
+  });
+}
 
 var corsOptions = {
   origin: "http://localhost:8081"
